remove useless interface
